% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVbolusPKmodels_Rfunctions.R
\name{OneCompIVbolus}
\alias{OneCompIVbolus}
\title{Process simulations using 1-compartment IV-bolus model.}
\usage{
OneCompIVbolus(inputDataFrame)
}
\arguments{
\item{inputDataFrame}{which is the NONMEM-style data frame that contains columns for \code{ID, TIME, AMT, CL, and V}.}
}
\value{
The function calculates the drug amount \code{(A1)} and individual predicted concentrations \code{(IPRED)}
in the central compartment and returns the output added to the \code{inputDataFrame}.
}
\description{
\code{OneCompIVbolus} function accepts NONMEM-style data frame for one subject and calculates drug amount in the respective compartments. The data frame should have the following columns: \code{ID, TIME, AMT, CL, V}.

where:
\tabular{ll}{
   \code{ID}: \tab is the subject ID\cr
   \code{TIME}:\tab is the sampling time points\cr
   \code{AMT}:\tab is the dose\cr
   \code{CL}:\tab is the central compartment clearance\cr
   \code{V}:\tab is the central volume of distribution\cr
   }
}
\details{
To simulate a population (i.e. the \code{inputDataFrame} has more than one subject \code{ID}),    the function
has to be applied for each subject \code{ID}. One way of doing that is through using the \code{ddply} functionality
in the \pkg{plyr} package in R. The \code{ddply} functionality allows applying the \pkg{PKADVAN} function to each subject
\code{ID} and combines the results into a data frame. Please load \pkg{plyr} package in \code{R} before
processsing simulations.

\code{ddply(inputDataFrame, .(ID), OneCompIVbolus)}

The \pkg{PKADVAN} function is capable of simulating arbitrary dosing regemins and can account for covariate structures; however,
covariate effects on respective parameters must be calculated prior porocessing simulations.
See examples below for more details.
}
\author{
Ahmad Abuhelwa\cr
Australian Center for Pharmacometrics\cr
School of Pharmacy and Medical Sciences\cr
University of South Australia\cr
\email{Ahmad.Abuhelwa@myamil.unisa.edu.au}
}
\seealso{
\code{\link{TwoCompIVbolus}}

\code{\link{ThreeCompIVbolus}}
}

